added patch selection via midi program change, separated note stack and midi cc code from midi.h
